==> 2 app data processing/2_selecting appropriate variables from raw data.ipynb <==
# raw data에서 필요한 column 고르기

#-*- coding: utf-8 -*-

import csv
import os
import sys

os.chdir('E:/apps')

### application_application.tsv

여기서 건질 column : number (app_id), granted (granted dummy)   
  
__problem__   
1. __앞자리가 0으로 시작하는 number__는 0이 빠져서 / 위치가 이상함   
=> 변환 필요  
__ex) 85/37802 -> 08/537802 __
2. number에 app_id와 DN이 섞여있음  
=> 따로 뽑자   


def app_id(aid):
    if len(aid) == 8:
        aid = aid.replace("/","")
        aid = '0' + aid[0] + '/' + aid[1:]
    
    return aid

with open('application_application.tsv','rU') as tsv, open('app_granted.csv','wb') as fw:
    reader = csv.reader(tsv, delimiter='\t', quoting=csv.QUOTE_NONE)
    reader.next()
    
    writer = csv.writer(fw)
    header = ['number','app_id','granted_dum']
    writer.writerow(header)  
    
    for line in reader:
        try:
            row=[]
            if len(line[2]) < 10:
                row.append('')
                row.append(app_id(line[2]))
            else:
                row.append(line[2])
                row.append('')
            
            row.append(line[7])
            writer.writerow(row)
            
        except IndexError:
            pass

__another problem__  
1. DN으로 나오는 애들 granted가 자꾸 null로 뜸  
=> 얘네만 따로 csv 만들어보자

with open('application_application.tsv','rU') as tsv, open('dn.csv','wb') as shit:
    reader = csv.reader(tsv, delimiter='\t', quoting=csv.QUOTE_NONE)
    header = reader.next()
    
    writer = csv.writer(shit)
    writer.writerow(header)
    
    for line in reader:
        if len(line[2]) > 10:
            writer.writerow(line)    

그냥 원래 null로 적혀있음 어쩔 수 없음

__app class 정리__

#-*- coding: utf-8 -*-

import csv
import os

os.chdir('E:/apps/app')

with open('app_class_raw.csv','rb') as f, open('app_class.csv','wb') as fw:
    reader = csv.reader(f)
    writer = csv.writer(fw)
    
    target = [0,2,3]
    
    for line in reader:
        try:
            writer.writerow([line[i] for i in target])
            
        except:
            continue
