==> 2 app data processing/4_app_cite + app_granted.ipynb <==
# app citation에 grant dummy 붙이기 

### merge 관련 issues  
1. 분석 범위: 2002년 ~ 2013년  
=> app_cite, app_granted 모두 2002 ~ 2013에 해당하는 친구들만 필요
2. 2002 ~ 2013에 해당하는 app은 app_profile에 있음
3. app_cite data가 아직 없어서 app_id 썼는지 DN 썼는지 모름  
=> 그렇다면 일단 app_profile에 granted dummy를 붙이자
4. 하지만 app_granted는 app_id, DN 섞여있음  
=> 2개 key (app_id, DN) 각각 join 후, column을 하나로 합침

### 1. app_profile + granted dummy

import csv
import os
import pandas as pd
import numpy as np
import math

os.chdir('E:/apps')

grt = pd.read_csv('app_granted.csv')
grt.iloc[0:5,:]

__문제__
- granted_dum에는 0, 1, NaN이 섞여 있음   
=> Join 시 생기는 NaN과 원래 있었던 NaN 구분 불가 
- 그렇다면 원래 있는 NaN을 2로 바꿔버리자!

grt['granted_dum'].fillna(2, inplace=True)
grt.loc[grt['number'] == 20050000001]

grt_id = grt.iloc[:,1:3]
grt_id.rename(columns={'granted_dum': 'granted_dum_i'}, inplace=True)

grt_n = grt.iloc[:,[0,2]]
grt_n.rename(columns={'granted_dum': 'granted_dum_n'}, inplace=True)

print list(grt_id), list(grt_n)

del grt

prf = pd.read_csv('app_profile_melted.csv')

prf = prf.merge(grt_id, how='left')
prf = prf.merge(grt_n, how='left')

merge 후 granted_dum_i, granted_dum_i 중 __하나만 NaN 일 때__ 정상  
둘 다 NaN일 경우 (matching 실패) or 둘 다 NaN 아닐 경우 (matching 중복) 처리 필요  
=> numpy array로 처리 후 다시 갖다 붙이자

prf_dum = prf.iloc[:,5:7]
prf.to_csv('app_prf_temp.csv',index=False)
prf_dum.to_csv('grtdum_temp.csv', index=False)

# dummy 뒷처리

def granted_dum(arr):
    a = arr[0] + arr[1]
    if math.isnan(a):
        if not math.isnan(arr[0]):
            a = arr[0]
        else:
            a = arr[1] if not math.isnan(arr[1]) else a
            a = np.nan if a==2 else a
     
    else:
        if a==2:
            if arr[0]==arr[1]:
                a=1
            
            else:
                a=0
        
        elif a==3:
            a=1
            
    return a

dum = np.genfromtxt('grtdum_temp.csv',delimiter=',',names=True)

vfunc = np.vectorize(granted_dum)
result = vfunc(dum)
np.savetxt('grtdum.csv', result,delimiter=',', fmt='%1.1f', header='granted_dum', comments='')

prf = pd.read_csv('app_prf_temp.csv')
dum = pd.read_csv('grtdum.csv')

prf = prf.iloc[:,0:5]

print len(prf), len(dum)

result = pd.concat([prf, dum], axis=1)
result.iloc[0:5,:]

result.to_csv('app_profile_granted.csv',index=False)
